requirements.txt
playwright
pandas
retry-async
python-json-logger
python-dotenv